previous node styles where not properly reset during rebuild
because the node's font was per default set to an object, the
refresh of the dataset (getRefreshedDataSet) could not clean the
old data since it was not null or undefined

(e.g. when updating the live view a node once was a
match and now is a neighbour and should not keep the match
style and vice versa).