Removing generated bundle from git versioning
Maybe felix will disagree with this decision, but people are asking for
a stable link to the bundle, and providing it through github's release
feature is the best way to do that, moreso than through an actual
repository directory.

I'm removing the version number from the generated source to create
a stable download link for #430.